Add test for getting apps; split helpers into own files
